cosomis/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.24.61
botocore==1.27.61
cachetools==4.2.4
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudant==2.15.0
decorator==4.4.2
Deprecated==1.2.14
Django==4.1.1
django-bootstrap4==22.2
django-celery-results==2.4.0
django-environ==0.9.0
django-htmx==1.17.3
django-storages==1.13.1
django-unicorn==0.48.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.2
drf-spectacular==0.23.1
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
Markdown==3.7
numpy==1.23.3
openpyxl==3.0.10
orjson==3.8.0
packaging==23.1
pandas==1.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
Pillow
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0.1
PyMySQL==1.1.0
redis==4.3.4
referencing==0.30.2
requests==2.28.1
rpds-py==0.10.3
s3transfer==0.6.2
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
xhtml2pdf==0.2.16
